agent-base/requirements.txt
agents-api==0.7.0.33
aio-pika==6.8.0
aiohttp==3.7.4.post0
aiormq==3.3.1
async-timeout==3.0.1
attrs==21.2.0
certifi==2021.5.30
chardet==4.0.0
charset-normalizer==2.0.4
data-source-api==0.4.0.23
googleapis-common-protos==1.53.0
grpcio==1.39.0
grpcio-tools==1.39.0
idna==3.2
lz4==3.1.3
multidict==5.1.0
pamqp==2.3.0
protobuf==3.17.3
python-consul2==0.1.5
PyYAML==5.4.1
requests==2.26.0
six==1.16.0
typing-extensions==3.10.0.0
urllib3==1.26.6
yarl==1.6.3
